fixed some bugs with downloading a runner, and moved tests back to each file.
