+++ src/runners/dxvk.rs
@@ -1,5 +1,5 @@
+use super::common::{AssetFilter, BaseGitHubRunner, GitHubRunnerConfig};
 use super::{Runner, RunnerManager, RunnerType};
-use super::common::{BaseGitHubRunner, GitHubRunnerConfig};
 use anyhow::{anyhow, Result};
 use regex::Regex;
 use std::path::{Path, PathBuf};
@@ -11,21 +11,21 @@ pub struct DxvkManager {
 
 impl DxvkManager {
     pub fn new(cellar_runners_path: PathBuf) -> Self {
+        fn asset_filter(name: &str) -> bool {
+            name.ends_with(".tar.gz") && !name.contains("source")
+        }
+
         let config = GitHubRunnerConfig {
             repo_owner: "doitsujin".to_string(),
             repo_name: "dxvk".to_string(),
             user_agent: "cellar/0.1.0".to_string(),
             max_download_size: 1024 * 1024 * 1024, // 1GB
-            max_files: 1000,
-            max_total_size: 2 * 1024 * 1024 * 1024, // 2GB
-            asset_filter: Box::new(|name: &str| name.ends_with(".tar.gz") && !name.contains("source")),
+            asset_filter: asset_filter as AssetFilter,
         };
-        
+
         let base_runner = BaseGitHubRunner::new(config, cellar_runners_path);
-        
-        Self {
-            base_runner,
-        }
+
+        Self { base_runner }
     }
 
     pub async fn discover_cellar_dxvk(&self) -> Result<Vec<Runner>> {
@@ -82,7 +82,61 @@ impl DxvkManager {
     }
 
     pub async fn extract_dxvk(&self, archive_path: &Path, version: &str) -> Result<PathBuf> {
-        self.base_runner.extract_runner_archive(archive_path, "dxvk", &format!("v{version}")).await
+        let dxvk_dir = self.base_runner.cellar_runners_path.join("dxvk");
+        fs::create_dir_all(&dxvk_dir).await?;
+
+        let extract_path = dxvk_dir.join(format!("v{version}"));
+        fs::create_dir_all(&extract_path).await?;
+
+        // Extract tar.gz file
+        let file = std::fs::File::open(archive_path)?;
+        let decoder = flate2::read::GzDecoder::new(file);
+        let mut archive = tar::Archive::new(decoder);
+
+        // Extract to temporary directory first
+        let temp_extract = std::env::temp_dir().join(format!("dxvk-extract-{version}"));
+        std::fs::create_dir_all(&temp_extract)?;
+        archive.unpack(&temp_extract)?;
+
+        // Find the extracted directory (usually the first subdirectory)
+        let mut entries = std::fs::read_dir(&temp_extract)?;
+        if let Some(entry) = entries.next() {
+            let extracted_dir = entry?.path();
+            if extracted_dir.is_dir() {
+                // Move contents to final destination
+                self.move_directory_contents(&extracted_dir, &extract_path)
+                    .await?;
+            }
+        }
+
+        // Clean up
+        std::fs::remove_dir_all(&temp_extract)?;
+        std::fs::remove_file(archive_path)?;
+
+        Ok(extract_path)
+    }
+
+    async fn move_directory_contents(&self, src: &Path, dest: &Path) -> Result<()> {
+        use std::collections::VecDeque;
+
+        let mut queue = VecDeque::new();
+        queue.push_back((src.to_path_buf(), dest.to_path_buf()));
+
+        while let Some((src_dir, dest_dir)) = queue.pop_front() {
+            let mut entries = fs::read_dir(&src_dir).await?;
+            while let Some(entry) = entries.next_entry().await? {
+                let src_path = entry.path();
+                let dest_path = dest_dir.join(entry.file_name());
+
+                if src_path.is_dir() {
+                    fs::create_dir_all(&dest_path).await?;
+                    queue.push_back((src_path, dest_path));
+                } else {
+                    fs::copy(&src_path, &dest_path).await?;
+                }
+            }
+        }
+        Ok(())
     }
 
     pub async fn install_dxvk_to_prefix(&self, dxvk_path: &Path, prefix_path: &Path) -> Result<()> {
