--- src/tests/mod.rs
@@ -1,5 +1,0 @@
-mod config;
-mod runners;
-mod utils;
-mod launch;
-mod cli;
